+++ src/app/(with-nav-footer)/activities/[id]/ActivityDetail/_components/ActivityReservation/PeopleCounter.tsx
@@ -1,0 +1,43 @@
+import Image from 'next/image';
+import decrement from '@/assets/icons/decrement.svg';
+import increment from '@/assets/icons/increment.svg';
+import { UseFormRegister, UseFormSetValue } from 'react-hook-form';
+
+type ReservationFormValues = {
+  date: string;
+  people: number;
+};
+
+type PeopleCounterProps = {
+  peopleCount: number;
+  setValue: UseFormSetValue<ReservationFormValues>;
+  register: UseFormRegister<ReservationFormValues>;
+};
+
+const PeopleCounter = ({ peopleCount, setValue, register }: PeopleCounterProps) => {
+  return (
+    <div className='flex w-[120px] items-center rounded-md border border-gray-300'>
+      <button
+        type='button'
+        className='w-[40px] p-2 text-center md:cursor-pointer'
+        onClick={() => setValue('people', Math.max(1, peopleCount - 1))}
+        aria-label='감소 버튼'
+      >
+        <Image src={decrement} width={18} height={18} alt='감소 버튼 아이콘' />
+      </button>
+
+      <input type='text' {...register('people', { required: true })} readOnly className='w-[40px] p-2 text-center' />
+
+      <button
+        type='button'
+        className='w-[40px] p-2 text-center md:cursor-pointer'
+        onClick={() => setValue('people', peopleCount + 1)}
+        aria-label='증가 버튼'
+      >
+        <Image src={increment} width={18} height={18} alt='증가 버튼 아이콘' />
+      </button>
+    </div>
+  );
+};
+
+export default PeopleCounter;
